refactor: image handling by introducing StoreBase64Image trait and update PDF generation trait
